# api/openapi/zke-server.v1.yaml
openapi: 3.1.0
info:
  title: ZKE Server API
  version: v1
  description: |
    ZKE 用户管理、资源管理、Cluster 接入管理和内部 Agent 注册使用的 HTTP API 契约。
    所有 JSON 时间均为 RFC 3339，并由当前 API 以 UTC+8 偏移返回。
servers:
  - url: /
    description: Console 与 API 使用同源入口；本地开发通过前端代理转发
...
paths:
...
  /api/v1/clusters/{cluster_id}:
    get:
      tags: [Resources]
      operationId: getCluster
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/ClusterID"
      responses:
        "200":
          description: Cluster 详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClusterAggregate"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
    put:
      tags: [Resources]
      operationId: updateCluster
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/ClusterID"
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClusterRequest"
      responses:
        "200":
          description: Cluster 名称已更新
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cluster"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      tags: [Resources]
      operationId: deleteCluster
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/ClusterID"
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Confirmation"
      responses:
        "200":
          description: Cluster 已退役，内部接入身份和凭证已撤销
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cluster"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: zke_session
    enrollmentToken:
      type: http
      scheme: bearer
      bearerFormat: Agent enrollment token
    installationToken:
      type: http
      scheme: bearer
      bearerFormat: Agent installation token
  parameters:
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
...
    ClusterID:
      name: cluster_id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/UUID"
...
  responses:
...
    Unauthenticated:
      description: 未认证或 Session 已失效
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: CSRF 或权限检查失败
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: 目标不存在或不可见
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: 幂等、生命周期或安全不变量冲突
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    UUID:
      type: string
      format: uuid
    Timestamp:
      type: string
      format: date-time
      examples:
        - "2026-07-26T14:30:00+08:00"
    Error:
      type: object
      additionalProperties: false
      required: [code, message, request_id]
      properties:
        code:
          type: string
        message:
          type: string
        request_id:
          type: string
...
    Confirmation:
      type: object
      additionalProperties: false
      required: [confirm]
      properties:
        confirm:
          const: true
...
    UpdateClusterRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 253
...
    ResourceBase:
      type: object
      required: [id, name, status, created_at, updated_at]
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        name:
          type: string
        status:
          type: string
        created_at:
          $ref: "#/components/schemas/Timestamp"
        updated_at:
          $ref: "#/components/schemas/Timestamp"
        replayed:
          type: boolean
...
    Cluster:
      allOf:
        - $ref: "#/components/schemas/ResourceBase"
        - type: object
          required: [tenant_id, project_id]
          properties:
            tenant_id:
              $ref: "#/components/schemas/UUID"
            project_id:
              $ref: "#/components/schemas/UUID"
            last_seen_at:
              $ref: "#/components/schemas/Timestamp"
    ClusterConnection:
      type: object
      additionalProperties: false
      required:
        - status
        - lifecycle_status
        - health_status
        - version
        - protocol_version
        - certificate_serial
        - certificate_expires_at
        - certificate_remaining_seconds
        - certificate_status
      properties:
        status:
          type: string
          enum: [online, offline]
        lifecycle_status:
          type: string
        health_status:
          type: string
        version:
          type: string
        protocol_version:
          type: string
        last_seen_at:
          $ref: "#/components/schemas/Timestamp"
        certificate_serial:
          type: string
        certificate_expires_at:
          $ref: "#/components/schemas/Timestamp"
        certificate_remaining_seconds:
          type: integer
          minimum: 0
        certificate_status:
          type: string
          enum: [valid, expiring, expired, revoked]
        connection_id:
          type: string
        connected_at:
          $ref: "#/components/schemas/Timestamp"
        last_heartbeat_at:
          $ref: "#/components/schemas/Timestamp"
        last_disconnected_at:
          $ref: "#/components/schemas/Timestamp"
        last_disconnect_reason:
          type: string
    ClusterAggregate:
      allOf:
        - $ref: "#/components/schemas/Cluster"
        - type: object
          required: [connection]
          properties:
            connection:
              $ref: "#/components/schemas/ClusterConnection"
...